Reshuffle benchmarks so different benches can use whatever setup is most convenient. Add embryonal Paillier benches
