+++ Cargo.toml
@@ -24,6 +24,7 @@ rand_chacha = { version = "0.3", default-features = false }
 rand = { version = "0.8", default-features = false }
 elliptic-curve = { version = "0.13", default-features = false }
 ecdsa = { version = "0.16", default-features = false, features = ["signing", "verifying"] }
+criterion = { version = "0.5", optional = true }
 
 # Note: `alloc` is needed for `crytpto-bigint`'s dependency `serdect` to be able
 # to serialize Uints in human-readable formats.
@@ -53,7 +54,7 @@ hex = { version = "0.4", default-features = false, features = ["alloc"] }
 test-log = { version = "0.2.16", default-features = false, features = ["trace", "color"] }
 
 [features]
-private-benches = ["k256"]
+private-benches = ["k256", "dep:criterion"]
 k256 = ["dep:k256", "bip32?/secp256k1"]
 bip32 = ["dep:bip32", "tiny-curve?/bip32"]
 dev = ["tiny-curve"]
@@ -67,5 +68,11 @@ required-features = ["dev"]
 [[bench]]
 bench = true
 name = "zk_proofs"
+harness = false
+required-features = ["private-benches"]
+
+[[bench]]
+bench = true
+name = "paillier"
 harness = false
 required-features = ["private-benches"]
